+++ app/jobs/send_admin_notification_job.rb
@@ -1,0 +1,7 @@
+class SendAdminNotificationJob < ApplicationJob
+  queue_as :background
+
+  def perform
+    AdminNotificationMailer.new_message_email.deliver_now
+  end
+end