+++ app/mailers/admin_notification_mailer.rb
@@ -1,0 +1,5 @@
+class AdminNotificationMailer < ApplicationMailer
+  def new_message_email
+    mail(to: "info@cbeebies-text.uk", subject: "New message received")
+  end
+end