+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_07_07_083843) do
+ActiveRecord::Schema[8.1].define(version: 2026_08_12_100913) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -50,6 +50,13 @@ ActiveRecord::Schema[8.1].define(version: 2026_07_07_083843) do
     t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
+  end
+
+  create_table "admin_notifications", force: :cascade do |t|
+    t.datetime "created_at", null: false
+    t.date "sent_on", null: false
+    t.datetime "updated_at", null: false
+    t.index ["sent_on"], name: "index_admin_notifications_on_sent_on", unique: true
   end
 
   create_table "admins", force: :cascade do |t|